+++ .prettierrc.js
@@ -4,4 +4,12 @@ module.exports = {
   printWidth: 100,
   singleQuote: true,
   trailingComma: 'es5',
+  overrides: [
+    {
+      files: ['*.ts', '*.tsx'],
+      options: {
+        parser: 'typescript',
+      },
+    },
+  ],
 };