+++ src/Pages/Accounts/Components/AccountButtons/Components/AccountAssets.js
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import styled from 'styled-components';
-import { useAccounts, useAssets } from 'redux/hooks';
+import { useAssets } from 'redux/hooks';
+import { useAccounts } from 'redux/new_hooks';
 import { Accordion, Table } from 'Components';
 
 const ButtonWrapper = styled.div`