+++ src/Pages/Accounts/Components/AccountSpotlight.js
@@ -1,7 +1,8 @@
 import React, { Fragment, useEffect, useState } from 'react';
 import styled from 'styled-components';
 import { useParams } from 'react-router-dom';
-import { useAccounts, useMediaQuery } from 'redux/hooks';
+import { useMediaQuery } from 'redux/hooks';
+import { useAccounts } from 'redux/new_hooks';
 import { Summary, Sprite, Loading } from 'Components';
 import { breakpoints } from 'consts';
 import { maxLength, formatDenom } from 'utils';