+++ src/redux/new_hooks/useAssets.ts
@@ -1,0 +1,18 @@
+import { useMemo } from 'react';
+import { bindActionCreators } from 'redux';
+import { useAppDispatch, useAppSelector } from 'redux/app/hooks';
+import {
+  selectAsset as selector,
+  assetActions as actionsList,
+} from 'redux/features/asset/assetSlice';
+
+export const useAssets = () => {
+  const dispatch = useAppDispatch();
+  const state = useAppSelector(selector);
+  const actions = useMemo(
+    () => bindActionCreators(actionsList, dispatch),
+    [dispatch]
+  );
+
+  return { ...state, ...actions };
+}