+++ src/theme/style.d.ts
@@ -1,15 +1,13 @@
 // import original module declarations
 import 'styled-components';
 import { string } from 'yup/lib/locale';
-import { Colors } from './Colors';
+import { Colors, COLORS_VALUE } from './Colors';
 import { Font } from './Font';
+import { DefaultThemeKeys } from './Themes';
 
 // and extend them!
 declare module 'styled-components' {
   interface DefaultTheme {
-    // To-Do: Update to populate intellisense. This
-    // just populates a list
-    [key: keyof Colors]: keyof typeof Colors,
-    [key: keyof Font]: keyof typeof Font,
+    [key in DefaultThemeKeys]?: COLORS_VALUE;
   }
 }