+++ src/utils/cookie/setCookie.ts
@@ -1,6 +1,9 @@
 import addDays from 'date-fns/addDays';
 
-export const setCookie = (cookieName, cookieValue, days = null) => {
+export const setCookie = (
+  cookieName: string, 
+  cookieValue: string, 
+  days?: number) => {
   // cookie method exists
   const expires = days ? `; expires=${addDays(new Date(), days)}` : '';
   const allowed = window && window.document && window.document.cookie !== undefined;